+++ src/audioatlas/__init__.py
@@ -21,4 +21,4 @@ __all__ = [
     "__version__",
 ]
 
-__version__ = "0.2.0a4"
+__version__ = "0.2.0a5"